+++ src/nzlusdb/suitability/indicators/__init__.py
@@ -1,0 +1,5 @@
+"""Python Code to compute land use indicators."""
+
+from nzlusdb import db
+
+INDICATORPATH = db.path / "indicators"