+++ CHANGELOG.md
@@ -5,4 +5,13 @@
 Contributor to this version: Baptiste Hamon (@baptistehamon).
 
 ### Announcements
-- The NZLUSDB website has been created ([GH#68](https://github.com/baptistehamon/lsapy/issues/68), [PR#74](https://github.com/baptistehamon/lsapy/pull/74)).
+- The NZLUSDB website has been created ([GH#4](https://github.com/baptistehamon/nzlusdb/issues/4), [PR#3](https://github.com/baptistehamon/nzlusdb/pull/3)).
+
+### Land Uses
+- The following land use has been added:
+  - Manuka ([GH#2](https://github.com/baptistehamon/nzlusdb/issues/2), [PR#8](https://github.com/baptistehamon/nzlusdb/pull/8))
+  - Citrus ([GH9#](https://github.com/baptistehamon/nzlusdb/issues/9), [PR#13](https://github.com/baptistehamon/nzlusdb/pull/13))
+
+### Internal Changes
+- A preprocessing step has been added enabling to preprocess criteria indicators before running the LSA ([GH9#](https://github.com/baptistehamon/nzlusdb/issues/9)).
+- The Command Line Interface (CLI) has been improved ([GH9#](https://github.com/baptistehamon/nzlusdb/issues/9)).